src/inventory-service/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.24.0
redis==5.0.1
# Note: Using psycopg2-binary is preferred for demos as it avoids needing gcc/pg_config in the container build
psycopg2-binary==2.9.9
boto3==1.29.6
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
opentelemetry-instrumentation-boto3sqs==0.42b0
prometheus-fastapi-instrumentator==7.1.0